Flip image so it acts like a mirror
Previously, it was disorienting to move around in the image since it was
acting like it was looking at you. A movement left would move right on
the screen. This change makes it act like a mirror so that a movement
left in real life acts like a movement left on screen.