Dilate and erode the binary image of the hand
In order to reduce the number of accidental holes in the hand due to
background subtraction, we want to dilate then erode the image to fill
in those smaller holes.

It might be useful in the future to erode first, then dilate twice, then
erode again. This would also remove any extraneous binary pieces due to
similar reasoning.